Support for remote raw commands (client/server)
See the '.:' command
Full integration with awk bindings
Add install make target for the awk bindings
Handle as debug mode if io->file->debug
Fix typo quite/quiet in rabin2
Resolve $pc,$bp,$sp,$a0.. in r_core_num
Minor random fixups

+++ libr/io/p/io_debug.c
@@ -300,7 +300,7 @@ struct r_io_plugin_t r_io_plugin_debug = {
         .plugin_open = __plugin_open,
 	.lseek = NULL,
 	.system = NULL,
-	.debug = (void *)1,
+	.debug = (void *)(size_t)1,
         //void *widget;
 /*
         struct debug_t *debug;
